move HeaderBlock to components/layout
