+++ src/components/Modals/AssetSelectorModal/AssetSelectorModal.js
@@ -25,7 +25,7 @@ import { useTranslationWithPrefix } from 'translations/translate';
 
 // Components
 import { Container } from 'components/layout/Layout';
-import HeaderBlock from 'components/HeaderBlock';
+import HeaderBlock from 'components/layout/HeaderBlock';
 import SlideModal from 'components/Modals/SlideModal';
 
 // Utils
