+++ src/components/Modals/ChainAssetSelectorModal/ChainAssetSelectorModal.js
@@ -27,7 +27,7 @@ import { useTranslationWithPrefix } from 'translations/translate';
 // Components
 import { Container } from 'components/layout/Layout';
 import TabView from 'components/layout/TabView';
-import HeaderBlock from 'components/HeaderBlock';
+import HeaderBlock from 'components/layout/HeaderBlock';
 import SlideModal from 'components/Modals/SlideModal';
 
 // Constants
