+++ src/screens/Asset/Asset.js
@@ -26,7 +26,7 @@ import { useNavigation, useNavigationParam } from 'react-navigation-hooks';
 
 // Components
 import { Container, Content } from 'components/layout/Layout';
-import HeaderBlock from 'components/HeaderBlock';
+import HeaderBlock from 'components/layout/HeaderBlock';
 import FloatingButtons from 'components/FloatingButtons';
 import ActivityFeed from 'components/legacy/ActivityFeed';
 import HistoryList from 'components/HistoryList';
