+++ src/screens/History/History.js
@@ -24,7 +24,7 @@ import { useTranslationWithPrefix } from 'translations/translate';
 
 // Components
 import { Container } from 'components/layout/Layout';
-import HeaderBlock from 'components/HeaderBlock';
+import HeaderBlock from 'components/layout/HeaderBlock';
 
 // Selectors
 import { useActiveAccount } from 'selectors';
