+++ src/screens/Home/Home.js
@@ -30,7 +30,7 @@ import { refreshEtherspotAccountsAction } from 'actions/etherspotActions';
 // Components
 import { Container, Content } from 'components/layout/Layout';
 import FloatingButtons from 'components/FloatingButtons';
-import HeaderBlock from 'components/HeaderBlock';
+import HeaderBlock from 'components/layout/HeaderBlock';
 import RefreshControl from 'components/RefreshControl';
 import Stories from 'components/Stories';
 import UserNameAndImage from 'components/UserNameAndImage';
