+++ src/screens/NewProfile/NewProfile.js
@@ -24,7 +24,7 @@ import t from 'translations/translate';
 
 // Components
 import { Container, Content } from 'components/layout/Layout';
-import HeaderBlock from 'components/HeaderBlock';
+import HeaderBlock from 'components/layout/HeaderBlock';
 import { Wrapper, Spacing } from 'components/legacy/Layout';
 import { BaseText, MediumText, Paragraph } from 'components/legacy/Typography';
 import Button from 'components/legacy/Button';
