+++ src/screens/SendToken/SendTokenConfirm.js
@@ -28,7 +28,7 @@ import { CHAIN } from 'constants/chainConstants';
 
 // components
 import { Container, Content, Spacing } from 'components/layout/Layout';
-import HeaderBlock from 'components/HeaderBlock';
+import HeaderBlock from 'components/layout/HeaderBlock';
 import Table, { TableRow, TableLabel, TableAmount, TableTotal, TableUser, TableFee } from 'components/legacy/Table';
 import Button from 'components/legacy/Button';
 import TokenReviewSummary from 'components/ReviewSummary/TokenReviewSummary';
